+++ packages/spacecat-shared-data-access/src/models/index.d.ts
@@ -10,6 +10,7 @@
  * governing permissions and limitations under the License.
  */
 
+export type * from './access-grant-log';
 export type * from './api-key';
 export type * from './async-job';
 export type * from './audit';
@@ -40,6 +41,7 @@ export type * from './sentiment-topic';
 export type * from './site';
 export type * from './site-candidate';
 export type * from './site-enrollment';
+export type * from './site-ims-org-access';
 export type * from './site-top-form';
 export type * from './site-top-page';
 export type * from './suggestion';