+++ packages/spacecat-shared-ims-client/CHANGELOG.md
@@ -1,3 +1,15 @@
+## [@adobe/spacecat-shared-ims-client-v1.12.1](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.12.0...@adobe/spacecat-shared-ims-client-v1.12.1) (2026-03-21)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1456](https://github.com/adobe/spacecat-shared/issues/1456)) ([b4654ec](https://github.com/adobe/spacecat-shared/commit/b4654ec88780f540f608fd8588182da805f2e6e2))
+
+## [@adobe/spacecat-shared-ims-client-v1.12.0](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.11.13...@adobe/spacecat-shared-ims-client-v1.12.0) (2026-03-16)
+
+### Features
+
+* add retry resilience to imsApiCall ([#1438](https://github.com/adobe/spacecat-shared/issues/1438)) ([2be8308](https://github.com/adobe/spacecat-shared/commit/2be8308ff6f308aa9f4bf75b0138f7a5dcefd7c3))
+
 ## [@adobe/spacecat-shared-ims-client-v1.11.13](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ims-client-v1.11.12...@adobe/spacecat-shared-ims-client-v1.11.13) (2026-03-15)
 
 ### Bug Fixes